project/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
django = "4.1.4"
pytest = "7.2.0"
pytest-django = "4.5.2"
celery = "5.2.7"
redis = "4.4.0"
flower = "1.2.0"
djangorestframework = "3.14.0"
pytelegrambotapi = "4.11.0"
psycopg2-binary = "2.9.6"
loguru = "^0.7.0"
gunicorn = "^20.1.0"
uuid = "^1.30"
timezonefinder = "^6.2.0"
acquirers = "^0.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
